Moved into individual folders and update build
